Revert "Moving from private inner class to public outer class to test"
This reverts commit 5fffc66b9ebfba6eb3ed207f45e2db1e297c5387.
